core: Moved render utils into own module
